+++ pymilo/pymilo_obj.py
@@ -2,12 +2,13 @@
 """PyMilo modules."""
 from .pymilo_func import get_sklearn_data, get_sklearn_version, to_sklearn_model
 from .utils.util import get_sklearn_type
-from .pymilo_param import PYMILO_VERSION
+from .pymilo_param import PYMILO_VERSION, PYMILO_VERSION_DOES_NOT_EXIST, UNEQUAL_PYMILO_VERSIONS
 import json
 
-from pymilo.exceptions.deserialize_exception import PymiloDeserializationException, DeSerilaizatoinErrorTypes
-from pymilo.exceptions.serialize_exception import PymiloSerializationException, SerilaizatoinErrorTypes
+from .exceptions.deserialize_exception import PymiloDeserializationException, DeSerilaizatoinErrorTypes
+from .exceptions.serialize_exception import PymiloSerializationException, SerilaizatoinErrorTypes
 from traceback import format_exc
+from warnings import warn
 
 
 class Export:
@@ -99,6 +100,10 @@ class Import:
             else:
                 with open(file_adr, 'r') as fp:
                     serialized_model_obj = json.load(fp)
+            if "pymilo_version" not in serialized_model_obj:
+                raise Exception(PYMILO_VERSION_DOES_NOT_EXIST)
+            if not serialized_model_obj["pymilo_version"] == PYMILO_VERSION:
+                warn(UNEQUAL_PYMILO_VERSIONS, category=Warning)
             self.data = serialized_model_obj["data"]
             self.version = serialized_model_obj["sklearn_version"]
             self.type = serialized_model_obj["model_type"]