+++ pymilo/utils/util.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 """utility module."""
-from numpy import ndarray
 from inspect import signature
 import importlib
 
@@ -45,21 +44,18 @@ def is_iterable(obj):
 
 def check_str_in_iterable(field, content):
     """
-    Check if the specified string field exists in content, which is supposed to be an iterable object.
+    Check if the specified string field exists in content, which is supposed to be a dictionary.
 
     :param field: given string field
     :type field: str
-    :param content: given supposed to be an iterable object
+    :param content: given supposed to be a dictionary
     :type content: obj
     :return: True if associated field is an iterable string in content and False otherwise.
     """
-    if not is_iterable(content):
-        return False
-    if isinstance(content, ndarray):
-        # https://stackoverflow.com/questions/40659212/futurewarning-elementwise-comparison-failed-returning-scalar-but-in-the-futur.
-        return False
-    else:
+    if isinstance(content, dict):
         return field in content
+    else:
+        return False
 
 
 def get_homogeneous_type(seq):
@@ -73,12 +69,10 @@ def get_homogeneous_type(seq):
     """
     iseq = iter(seq)
     first_type = type(next(iseq))
-    return (
-        True,
-        first_type) if all(
-        (type(x) is first_type) for x in iseq) else (
-            False,
-        None)
+    if all((isinstance(x, first_type)) for x in iseq):
+        return True, first_type
+    else:
+        return False, None
 
 
 def all_same(arr):